+++ lib/nsrr/helpers/constants.rb
@@ -1,0 +1,1 @@
+Nsrr::WEBSITE = "https://sleepdata.org"