+++ lib/nsrr/models/all.rb
@@ -1,0 +1,1 @@
+require 'nsrr/models/dataset'