+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.1.0
 
 ### Enhancements
+- Added a `nsrr console` command that allows users to access and download datasets and files
 - Added a `nsrr version` command the returns the current version of the nsrr gem
 - Added testing framework to more easily add new tests for new features